5. Обработка исключений и Spring REST [#504797]

+++ src/main/java/ru/job4j/controller/MessageController.java
@@ -3,6 +3,7 @@ package ru.job4j.controller;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import ru.job4j.domain.Message;
 import ru.job4j.service.ChatService;
 
@@ -24,15 +25,23 @@ public class MessageController {
 
     @GetMapping("/{id}")
     public ResponseEntity<Message> findById(@PathVariable int id) {
-        var message = this.service.findMessageById(id);
+        if (id <= 0) {
+            throw new NullPointerException("Id should be more than 0");
+        }
+        Message message = this.service.findMessageById(id).orElseThrow(() -> new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Message not found. Please, check id."
+        ));
         return new ResponseEntity<>(
-                message.orElse(new Message()),
-                message.isPresent() ? HttpStatus.OK : HttpStatus.NOT_FOUND
+                message,
+                HttpStatus.OK
         );
     }
 
     @PostMapping("/")
     public ResponseEntity<Message> create(@RequestBody Message message) {
+        if (message.getText() == null) {
+            throw new NullPointerException("Message text can`t be empty");
+        }
         return new ResponseEntity<>(
                 this.service.saveMessage(message),
                 HttpStatus.CREATED
@@ -41,12 +50,18 @@ public class MessageController {
 
     @PutMapping("/")
     public ResponseEntity<Void> update(@RequestBody Message message) {
+        if (message.getText() == null) {
+            throw new NullPointerException("Message text can`t be empty");
+        }
         this.service.saveMessage(message);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable int id) {
+        if (id <= 0) {
+            throw new NullPointerException("Id should be more than 0");
+        }
         Message message = new Message();
         message.setId(id);
         this.service.deleteMessage(message);
