+++ src/main/java/ru/job4j/controller/RoomController.java
@@ -4,6 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.client.RestTemplate;
+import org.springframework.web.server.ResponseStatusException;
 import ru.job4j.domain.Message;
 import ru.job4j.domain.Room;
 import ru.job4j.service.ChatService;
@@ -29,15 +30,23 @@ public class RoomController {
 
     @GetMapping("/{id}")
     public ResponseEntity<Room> findById(@PathVariable int id) {
-        var room = this.service.findRoomById(id);
+        if (id <= 0) {
+            throw new NullPointerException("Id should be more than 0");
+        }
+        Room room = this.service.findRoomById(id).orElseThrow(() -> new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Room not found. Please, check id."
+        ));
         return new ResponseEntity<>(
-                room.orElse(new Room()),
-                room.isPresent() ? HttpStatus.OK : HttpStatus.NOT_FOUND
+                room,
+                HttpStatus.OK
         );
     }
 
     @PostMapping("/")
     public ResponseEntity<Room> create(@RequestBody Room room) {
+        if (room.getName() == null) {
+            throw new NullPointerException("Room name can`t be empty");
+        }
         return new ResponseEntity<>(
                 this.service.saveRoom(room),
                 HttpStatus.CREATED
@@ -46,12 +55,18 @@ public class RoomController {
 
     @PutMapping("/")
     public ResponseEntity<Void> update(@RequestBody Room room) {
+        if (room.getName() == null) {
+            throw new NullPointerException("Room name can`t be empty");
+        }
         this.service.saveRoom(room);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable int id) {
+        if (id <= 0) {
+            throw new NullPointerException("Id should be more than 0");
+        }
         Room room = new Room();
         room.setId(id);
         this.service.deleteRoom(room);
@@ -60,6 +75,12 @@ public class RoomController {
 
     @PutMapping("/{roomId}/addMessage")
     public ResponseEntity<Void> addMessageToRoom(@PathVariable int roomId, @RequestBody Message message) {
+        if (roomId <= 0) {
+            throw new NullPointerException("roomId should be more than 0");
+        }
+        if (message.getText() == null) {
+            throw new NullPointerException("Message text can`t be empty");
+        }
         Room room = findById(roomId).getBody();
         Message savedMessage = service.saveMessage(message);
         room.addMessage(savedMessage);
